Improve email delivery service logic for thirdpartypasswordless recipe
