+++ supertokens_python/recipe/passwordless/types.py
@@ -13,7 +13,6 @@
 # under the License.
 from supertokens_python.ingredients.emaildelivery import \
     EmailDeliveryIngredient
-from .interfaces import TypePasswordlessEmailDeliveryInput
 from typing import List, Union
 
 
@@ -46,6 +45,23 @@ class DeviceType:
         self.codes = codes
         self.email = email
         self.phone_number = phone_number
+
+
+class CreateAndSendCustomEmailParameters:
+    def __init__(self,
+                 code_life_time: int,
+                 pre_auth_session_id: str,
+                 email: str,
+                 user_input_code: Union[str, None] = None,
+                 url_with_link_code: Union[str, None] = None):
+        self.email: str = email
+        self.code_life_time: int = code_life_time
+        self.pre_auth_session_id: str = pre_auth_session_id
+        self.user_input_code: Union[str, None] = user_input_code
+        self.url_with_link_code: Union[str, None] = url_with_link_code
+
+
+TypePasswordlessEmailDeliveryInput = CreateAndSendCustomEmailParameters
 
 
 class PasswordlessIngredients:
